Moved storage helpers into dedicated file
